requirements.txt
streamlit==1.31.0
langchain==0.1.0
langchain-community==0.0.13
openai==1.10.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
requests==2.31.0
beautifulsoup4==4.12.2
watchdog==3.0.0
setuptools>=65.5.1
wheel>=0.38.0